requirements.txt
py-cord
Pillow
requests
io